Move UI logic into separate `ui` package.
